auth module and server.js refactored to allow using of multiple knex objects
