+++ metaspace/graphql/server.js
@@ -13,10 +13,10 @@ const bodyParser = require('body-parser'),
   {promisify} = require('util'),
   readFile = promisify(require("fs").readFile);
 
-const {createImgServerAsync} = require('./imageUpload.js'),
-  {configureAuth, initSchema} = require('./src/modules/auth'),
+const { createImgServerAsync } = require('./imageUpload.js'),
+  { configureAuth } = require('./src/modules/auth'),
   Resolvers = require('./resolvers'),
-  {logger, initDBConnection} = require('./utils');
+  { logger, initDBConnection } = require('./utils');
 
 // subscriptions setup
 const http = require('http'),
@@ -46,83 +46,80 @@ const configureSession = (app) => {
   }));
 };
 
-function createHttpServerAsync(config) {
+async function createHttpServerAsync(config) {
   let app = express();
   let httpServer = http.createServer(app);
 
-  return initSchema()
-    .then(async () => {
-      return mergeTypes([
+  const mergedSchema = mergeTypes([
         await readFile('schema.graphql', 'utf8'),
         await readFile('schemas/user.graphql', 'utf8'),
-        await readFile('schemas/group.graphql', 'utf8'),
-      ]);
-    })
-    .then((mergedSchema) => {
-      const schema = makeExecutableSchema({typeDefs: mergedSchema});
-      addResolveFunctionsToSchema(schema, Resolvers);
-      addErrorLoggingToSchema(schema, logger);
+        await readFile('schemas/group.graphql', 'utf8')]);
 
-      if (process.env.NODE_ENV === 'development') {
-        addMockFunctionsToSchema({schema, preserveResolvers: true});
-      } else {
-        maskErrors(schema);
-      }
+  const schema = makeExecutableSchema({typeDefs: mergedSchema});
+  addResolveFunctionsToSchema(schema, Resolvers);
+  addErrorLoggingToSchema(schema, logger);
 
-      app.use(cors());
-      app.use(compression());
-      app.use(jwt({
-        secret: config.jwt.secret,
-        // issuer: config.jwt.issuer, // TODO: Add issuer to config so that it can be validated
-        credentialsRequired: false,
-      }));
-      app.use('/graphql',
-          bodyParser.json({type: '*/*'}),
-          graphqlExpress(req => ({
-            schema,
-            context: req
-          })));
-      app.use('/graphiql', graphiqlExpress({
-        endpointURL: '/graphql',
-        subscriptionsEndpoint: config.websocket_public_url,
-      }));
+  if (process.env.NODE_ENV === 'development') {
+    addMockFunctionsToSchema({schema, preserveResolvers: true});
+  } else {
+    maskErrors(schema);
+  }
 
-      if (config.features.newAuth) {
-        app.use(bodyParser.json());
-        configureSession(app);
-        configureAuth(app);
-      }
+  app.use(cors());
+  app.use(compression());
+  app.use(jwt({
+    secret: config.jwt.secret,
+    // issuer: config.jwt.issuer, // TODO: Add issuer to config so that it can be validated
+    credentialsRequired: false,
+  }));
+  app.use('/graphql',
+      bodyParser.json({type: '*/*'}),
+      graphqlExpress(req => ({
+        schema,
+        context: req
+      })));
+  app.use('/graphiql', graphiqlExpress({
+    endpointURL: '/graphql',
+    subscriptionsEndpoint: config.websocket_public_url,
+  }));
 
-      app.use(function (err, req, res, next) {
-        res.status(err.status || 500);
-        logger.error(err.stack);
-        res.json({
-          message: err.message
-        });
-      });
+  if (config.features.newAuth) {
+    app.use(bodyParser.json());
+    configureSession(app);
+    await configureAuth(app);
+  }
 
+  app.use(function (err, req, res, next) {
+    res.status(err.status || 500);
+    logger.error(err.stack);
+    res.json({
+      message: err.message
+    });
+  });
 
-      httpServer.listen(config.port);
+  httpServer.listen(config.port);
 
-      wsServer.listen(config.ws_port, (err) => {
-        if (err) {
-          logger.error('Could not start WebSocket server', err)
-        }
-        logger.info(`WebSocket server is running on ${config.ws_port} port...`);
-        SubscriptionServer.create({execute, subscribe, schema}, {
-          server: wsServer,
-          path: '/graphql',
-        });
-      });
+  wsServer.listen(config.ws_port, (err) => {
+    if (err) {
+      logger.error('Could not start WebSocket server', err)
+    }
+    logger.info(`WebSocket server is running on ${config.ws_port} port...`);
+    SubscriptionServer.create({execute, subscribe, schema}, {
+      server: wsServer,
+      path: '/graphql',
+    });
+  });
 
-      logger.info(`SM GraphQL is running on ${config.port} port...`);
+  logger.info(`SM GraphQL is running on ${config.port} port...`);
 
-      return httpServer;
-    })
+  return httpServer;
 }
 
 if (process.argv[1].endsWith('server.js')) {
-  createHttpServerAsync(config);
+  createHttpServerAsync(config)
+    .catch(e => {
+      logger.error(e);
+    });
   createImgServerAsync(config, initDBConnection());
 }
