+++ metaspace/graphql/src/modules/auth/index.ts
@@ -1,2 +1,2 @@
-export {configureAuth} from './api';
-export {initSchema} from './db';
+export {configureAuth} from './controller';
+export {initOperation} from './operation';
